requirements.txt
transformers==4.51.3
liger_kernel==0.5.10
torch==2.7.1
torchvision==0.22.1
torchaudio==2.7.1
deepspeed==0.17.0
accelerate==1.7.0
sed_eval==0.2.1
peft==0.17.1
hydra-core>=1.3.0
fire>=0.5.0
h5py>=3.10.0
ffmpeg>=1.4
decord>=0.6.0
trl<=0.19.0